Editando API

+++ app/src/main/java/uniandrade/br/edu/com/seriadosbrasileiros/SeriesResults.java
@@ -130,7 +130,7 @@ public class SeriesResults {
         private String media_type;
         private String name;
         private int vote_count;
-        private int vote_average;
+        private double vote_average;
         private String poster_path;
         private String first_air_date;
         private double popularity;
@@ -185,11 +185,11 @@ public class SeriesResults {
             this.vote_count = vote_count;
         }
 
-        public int getVote_average() {
+        public double getVote_average() {
             return vote_average;
         }
 
-        public void setVote_average(int vote_average) {
+        public void setVote_average(double vote_average) {
             this.vote_average = vote_average;
         }
